added

+++ os161-1.99/kern/syscall/proc_syscalls.c
@@ -148,8 +148,8 @@ sys_waitpid(pid_t pid,
     
     while (pinfo != NULL){
         if (pinfo->pid == pid){
-            kprintf("%d\n", pid);
-            kprintf("%d\n", pinfo->pid);
+            kprintf("%d\n", (int)pid);
+            kprintf("%d\n", (int)pinfo->pid);
             break;
         }
         pinfo = pinfo->next_sibling;
